src/main/java/com/learning/model/User.java
package com.learning.model;

public class User {
    private int id;
    private String loginId;
    private String password;
    public User() {}

    public User(String loginId, String password) {
        this.loginId = loginId;
        this.password = password;
    }

    public User(int id, String loginId, String password) {
    }

    public int getId() {
        return id;
    }

    public void setId(int id) {
        this.id = id;
    }

    public String getLoginId() {
        return loginId;
    }

    public void setLoginId(String loginId) {
        this.loginId = loginId;
    }

    public String getPassword() {
        return password;
    }

    public void setPassword(String password) {
        this.password = password;
    }
    @Override
    public String toString() {
        return "User{" +
                "id=" + id +
                ", loginId='" + loginId + '\'' +
                ", password='" + password + '\'' +
                '}';
    }
}
